+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added built-in support for the `BackgroundColor` bevy component, under the `bevy_ui` feature.
+
 ## [0.8.0] - 2023-07-12
 
 ### Changed